Move code from VirtualProjectBuildingCommand.cs to FileLevelDirectiveHelpers.cs
